sync-pipelines/requirements.txt
web3~=6.9.0
PyYAML~=6.0.1
peewee~=3.16.3
structlog~=23.1.0
psycopg2-binary~=2.9.7
